Refactor the datastructures code structure
Splitting the code into related modules makes it easier to maintain
the code.
