+++ src/werkzeug/datastructures/cache_control_headers.pyi
@@ -1,0 +1,114 @@
+from typing import Callable
+from typing import Dict
+from typing import Iterable
+from typing import Mapping
+from typing import Optional
+from typing import Tuple
+from typing import Type
+from typing import TypeVar
+from typing import Union
+
+from .mixins import ImmutableDictMixin
+from .mixins import UpdateDictMixin
+
+T = TypeVar("T")
+_CPT = TypeVar("_CPT", str, int, bool)
+_OptCPT = Optional[_CPT]
+
+def cache_control_property(key: str, empty: _OptCPT, type: Type[_CPT]) -> property: ...
+
+class _CacheControl(UpdateDictMixin[str, _OptCPT], Dict[str, _OptCPT]):
+    provided: bool
+    def __init__(
+        self,
+        values: Union[Mapping[str, _OptCPT], Iterable[Tuple[str, _OptCPT]]] = (),
+        on_update: Optional[Callable[[_CacheControl], None]] = None,
+    ) -> None: ...
+    @property
+    def no_cache(self) -> Optional[bool]: ...
+    @no_cache.setter
+    def no_cache(self, value: Optional[bool]) -> None: ...
+    @no_cache.deleter
+    def no_cache(self) -> None: ...
+    @property
+    def no_store(self) -> Optional[bool]: ...
+    @no_store.setter
+    def no_store(self, value: Optional[bool]) -> None: ...
+    @no_store.deleter
+    def no_store(self) -> None: ...
+    @property
+    def max_age(self) -> Optional[int]: ...
+    @max_age.setter
+    def max_age(self, value: Optional[int]) -> None: ...
+    @max_age.deleter
+    def max_age(self) -> None: ...
+    @property
+    def no_transform(self) -> Optional[bool]: ...
+    @no_transform.setter
+    def no_transform(self, value: Optional[bool]) -> None: ...
+    @no_transform.deleter
+    def no_transform(self) -> None: ...
+    def _get_cache_value(self, key: str, empty: Optional[T], type: Type[T]) -> T: ...
+    def _set_cache_value(self, key: str, value: Optional[T], type: Type[T]) -> None: ...
+    def _del_cache_value(self, key: str) -> None: ...
+    def to_header(self) -> str: ...
+    @staticmethod
+    def cache_property(key: str, empty: _OptCPT, type: Type[_CPT]) -> property: ...
+
+class RequestCacheControl(ImmutableDictMixin[str, _OptCPT], _CacheControl):
+    @property
+    def max_stale(self) -> Optional[int]: ...
+    @max_stale.setter
+    def max_stale(self, value: Optional[int]) -> None: ...
+    @max_stale.deleter
+    def max_stale(self) -> None: ...
+    @property
+    def min_fresh(self) -> Optional[int]: ...
+    @min_fresh.setter
+    def min_fresh(self, value: Optional[int]) -> None: ...
+    @min_fresh.deleter
+    def min_fresh(self) -> None: ...
+    @property
+    def only_if_cached(self) -> Optional[bool]: ...
+    @only_if_cached.setter
+    def only_if_cached(self, value: Optional[bool]) -> None: ...
+    @only_if_cached.deleter
+    def only_if_cached(self) -> None: ...
+
+class ResponseCacheControl(_CacheControl):
+    @property
+    def public(self) -> Optional[bool]: ...
+    @public.setter
+    def public(self, value: Optional[bool]) -> None: ...
+    @public.deleter
+    def public(self) -> None: ...
+    @property
+    def private(self) -> Optional[bool]: ...
+    @private.setter
+    def private(self, value: Optional[bool]) -> None: ...
+    @private.deleter
+    def private(self) -> None: ...
+    @property
+    def must_revalidate(self) -> Optional[bool]: ...
+    @must_revalidate.setter
+    def must_revalidate(self, value: Optional[bool]) -> None: ...
+    @must_revalidate.deleter
+    def must_revalidate(self) -> None: ...
+    @property
+    def proxy_revalidate(self) -> Optional[bool]: ...
+    @proxy_revalidate.setter
+    def proxy_revalidate(self, value: Optional[bool]) -> None: ...
+    @proxy_revalidate.deleter
+    def proxy_revalidate(self) -> None: ...
+    @property
+    def s_maxage(self) -> Optional[int]: ...
+    @s_maxage.setter
+    def s_maxage(self, value: Optional[int]) -> None: ...
+    @s_maxage.deleter
+    def s_maxage(self) -> None: ...
+    @property
+    def immutable(self) -> Optional[bool]: ...
+    @immutable.setter
+    def immutable(self, value: Optional[bool]) -> None: ...
+    @immutable.deleter
+    def immutable(self) -> None: ...
