+++ src/werkzeug/datastructures/headers.pyi
@@ -1,0 +1,312 @@
+from typing import Any
+from typing import Callable
+from typing import Collection
+from typing import Dict
+from typing import FrozenSet
+from typing import Iterable
+from typing import Iterator
+from typing import List
+from typing import Mapping
+from typing import NoReturn
+from typing import Optional
+from typing import overload
+from typing import Set
+from typing import Tuple
+from typing import Type
+from typing import TypeVar
+from typing import Union
+from _typeshed import SupportsKeysAndGetItem
+from _typeshed.wsgi import WSGIEnvironment
+
+from typing_extensions import Literal
+
+from .mixins import ImmutableHeadersMixin
+from .mixins import UpdateDictMixin
+
+D = TypeVar("D")
+T = TypeVar("T")
+HV = Union[str, int]
+
+class Headers(Dict[str, str]):
+    _list: List[Tuple[str, str]]
+    def __init__(
+        self,
+        defaults: Optional[
+            Union[Mapping[str, Union[HV, Iterable[HV]]], Iterable[Tuple[str, HV]]]
+        ] = None,
+    ) -> None: ...
+    @overload
+    def __getitem__(self, key: str) -> str: ...
+    @overload
+    def __getitem__(self, key: int) -> Tuple[str, str]: ...
+    @overload
+    def __getitem__(self, key: slice) -> Headers: ...
+    @overload
+    def __getitem__(self, key: str, _get_mode: Literal[True] = ...) -> str: ...
+    def __eq__(self, other: object) -> bool: ...
+    @overload  # type: ignore
+    def get(self, key: str, default: str) -> str: ...
+    @overload
+    def get(self, key: str, default: Optional[str] = None) -> Optional[str]: ...
+    @overload
+    def get(
+        self, key: str, default: Optional[T] = None, type: Callable[[str], T] = ...
+    ) -> Optional[T]: ...
+    @overload
+    def getlist(self, key: str) -> List[str]: ...
+    @overload
+    def getlist(self, key: str, type: Callable[[str], T]) -> List[T]: ...
+    def get_all(self, name: str) -> List[str]: ...
+    def items(  # type: ignore
+        self, lower: bool = False
+    ) -> Iterator[Tuple[str, str]]: ...
+    def keys(self, lower: bool = False) -> Iterator[str]: ...  # type: ignore
+    def values(self) -> Iterator[str]: ...  # type: ignore
+    def extend(
+        self,
+        *args: Union[Mapping[str, Union[HV, Iterable[HV]]], Iterable[Tuple[str, HV]]],
+        **kwargs: Union[HV, Iterable[HV]],
+    ) -> None: ...
+    @overload
+    def __delitem__(self, key: Union[str, int, slice]) -> None: ...
+    @overload
+    def __delitem__(self, key: str, _index_operation: Literal[False]) -> None: ...
+    def remove(self, key: str) -> None: ...
+    @overload  # type: ignore
+    def pop(self, key: str, default: Optional[str] = None) -> str: ...
+    @overload
+    def pop(
+        self, key: Optional[int] = None, default: Optional[Tuple[str, str]] = None
+    ) -> Tuple[str, str]: ...
+    def popitem(self) -> Tuple[str, str]: ...
+    def __contains__(self, key: str) -> bool: ...  # type: ignore
+    def has_key(self, key: str) -> bool: ...
+    def __iter__(self) -> Iterator[Tuple[str, str]]: ...  # type: ignore
+    def add(self, _key: str, _value: HV, **kw: HV) -> None: ...
+    def _validate_value(self, value: str) -> None: ...
+    def add_header(self, _key: str, _value: HV, **_kw: HV) -> None: ...
+    def clear(self) -> None: ...
+    def set(self, _key: str, _value: HV, **kw: HV) -> None: ...
+    def setlist(self, key: str, values: Iterable[HV]) -> None: ...
+    def setdefault(self, key: str, default: HV) -> str: ...  # type: ignore
+    def setlistdefault(self, key: str, default: Iterable[HV]) -> None: ...
+    @overload
+    def __setitem__(self, key: str, value: HV) -> None: ...
+    @overload
+    def __setitem__(self, key: int, value: Tuple[str, HV]) -> None: ...
+    @overload
+    def __setitem__(self, key: slice, value: Iterable[Tuple[str, HV]]) -> None: ...
+    @overload
+    def update(
+        self, __m: SupportsKeysAndGetItem[str, HV], **kwargs: Union[HV, Iterable[HV]]
+    ) -> None: ...
+    @overload
+    def update(
+        self, __m: Iterable[Tuple[str, HV]], **kwargs: Union[HV, Iterable[HV]]
+    ) -> None: ...
+    @overload
+    def update(self, **kwargs: Union[HV, Iterable[HV]]) -> None: ...
+    def to_wsgi_list(self) -> List[Tuple[str, str]]: ...
+    def copy(self) -> Headers: ...
+    def __copy__(self) -> Headers: ...
+
+class EnvironHeaders(ImmutableHeadersMixin, Headers):
+    environ: WSGIEnvironment
+    def __init__(self, environ: WSGIEnvironment) -> None: ...
+    def __eq__(self, other: object) -> bool: ...
+    def __getitem__(  # type: ignore
+        self, key: str, _get_mode: Literal[False] = False
+    ) -> str: ...
+    def __iter__(self) -> Iterator[Tuple[str, str]]: ...  # type: ignore
+    def copy(self) -> NoReturn: ...
+
+class ETags(Collection[str]):
+    _strong: FrozenSet[str]
+    _weak: FrozenSet[str]
+    star_tag: bool
+    def __init__(
+        self,
+        strong_etags: Optional[Iterable[str]] = None,
+        weak_etags: Optional[Iterable[str]] = None,
+        star_tag: bool = False,
+    ) -> None: ...
+    def as_set(self, include_weak: bool = False) -> Set[str]: ...
+    def is_weak(self, etag: str) -> bool: ...
+    def is_strong(self, etag: str) -> bool: ...
+    def contains_weak(self, etag: str) -> bool: ...
+    def contains(self, etag: str) -> bool: ...
+    def contains_raw(self, etag: str) -> bool: ...
+    def to_header(self) -> str: ...
+    def __call__(
+        self,
+        etag: Optional[str] = None,
+        data: Optional[bytes] = None,
+        include_weak: bool = False,
+    ) -> bool: ...
+    def __len__(self) -> int: ...
+    def __iter__(self) -> Iterator[str]: ...
+    def __contains__(self, item: str) -> bool: ...  # type: ignore
+
+def csp_property(key: str) -> property: ...
+
+class ContentSecurityPolicy(UpdateDictMixin[str, str], Dict[str, str]):
+    @property
+    def base_uri(self) -> Optional[str]: ...
+    @base_uri.setter
+    def base_uri(self, value: Optional[str]) -> None: ...
+    @base_uri.deleter
+    def base_uri(self) -> None: ...
+    @property
+    def child_src(self) -> Optional[str]: ...
+    @child_src.setter
+    def child_src(self, value: Optional[str]) -> None: ...
+    @child_src.deleter
+    def child_src(self) -> None: ...
+    @property
+    def connect_src(self) -> Optional[str]: ...
+    @connect_src.setter
+    def connect_src(self, value: Optional[str]) -> None: ...
+    @connect_src.deleter
+    def connect_src(self) -> None: ...
+    @property
+    def default_src(self) -> Optional[str]: ...
+    @default_src.setter
+    def default_src(self, value: Optional[str]) -> None: ...
+    @default_src.deleter
+    def default_src(self) -> None: ...
+    @property
+    def font_src(self) -> Optional[str]: ...
+    @font_src.setter
+    def font_src(self, value: Optional[str]) -> None: ...
+    @font_src.deleter
+    def font_src(self) -> None: ...
+    @property
+    def form_action(self) -> Optional[str]: ...
+    @form_action.setter
+    def form_action(self, value: Optional[str]) -> None: ...
+    @form_action.deleter
+    def form_action(self) -> None: ...
+    @property
+    def frame_ancestors(self) -> Optional[str]: ...
+    @frame_ancestors.setter
+    def frame_ancestors(self, value: Optional[str]) -> None: ...
+    @frame_ancestors.deleter
+    def frame_ancestors(self) -> None: ...
+    @property
+    def frame_src(self) -> Optional[str]: ...
+    @frame_src.setter
+    def frame_src(self, value: Optional[str]) -> None: ...
+    @frame_src.deleter
+    def frame_src(self) -> None: ...
+    @property
+    def img_src(self) -> Optional[str]: ...
+    @img_src.setter
+    def img_src(self, value: Optional[str]) -> None: ...
+    @img_src.deleter
+    def img_src(self) -> None: ...
+    @property
+    def manifest_src(self) -> Optional[str]: ...
+    @manifest_src.setter
+    def manifest_src(self, value: Optional[str]) -> None: ...
+    @manifest_src.deleter
+    def manifest_src(self) -> None: ...
+    @property
+    def media_src(self) -> Optional[str]: ...
+    @media_src.setter
+    def media_src(self, value: Optional[str]) -> None: ...
+    @media_src.deleter
+    def media_src(self) -> None: ...
+    @property
+    def navigate_to(self) -> Optional[str]: ...
+    @navigate_to.setter
+    def navigate_to(self, value: Optional[str]) -> None: ...
+    @navigate_to.deleter
+    def navigate_to(self) -> None: ...
+    @property
+    def object_src(self) -> Optional[str]: ...
+    @object_src.setter
+    def object_src(self, value: Optional[str]) -> None: ...
+    @object_src.deleter
+    def object_src(self) -> None: ...
+    @property
+    def prefetch_src(self) -> Optional[str]: ...
+    @prefetch_src.setter
+    def prefetch_src(self, value: Optional[str]) -> None: ...
+    @prefetch_src.deleter
+    def prefetch_src(self) -> None: ...
+    @property
+    def plugin_types(self) -> Optional[str]: ...
+    @plugin_types.setter
+    def plugin_types(self, value: Optional[str]) -> None: ...
+    @plugin_types.deleter
+    def plugin_types(self) -> None: ...
+    @property
+    def report_to(self) -> Optional[str]: ...
+    @report_to.setter
+    def report_to(self, value: Optional[str]) -> None: ...
+    @report_to.deleter
+    def report_to(self) -> None: ...
+    @property
+    def report_uri(self) -> Optional[str]: ...
+    @report_uri.setter
+    def report_uri(self, value: Optional[str]) -> None: ...
+    @report_uri.deleter
+    def report_uri(self) -> None: ...
+    @property
+    def sandbox(self) -> Optional[str]: ...
+    @sandbox.setter
+    def sandbox(self, value: Optional[str]) -> None: ...
+    @sandbox.deleter
+    def sandbox(self) -> None: ...
+    @property
+    def script_src(self) -> Optional[str]: ...
+    @script_src.setter
+    def script_src(self, value: Optional[str]) -> None: ...
+    @script_src.deleter
+    def script_src(self) -> None: ...
+    @property
+    def script_src_attr(self) -> Optional[str]: ...
+    @script_src_attr.setter
+    def script_src_attr(self, value: Optional[str]) -> None: ...
+    @script_src_attr.deleter
+    def script_src_attr(self) -> None: ...
+    @property
+    def script_src_elem(self) -> Optional[str]: ...
+    @script_src_elem.setter
+    def script_src_elem(self, value: Optional[str]) -> None: ...
+    @script_src_elem.deleter
+    def script_src_elem(self) -> None: ...
+    @property
+    def style_src(self) -> Optional[str]: ...
+    @style_src.setter
+    def style_src(self, value: Optional[str]) -> None: ...
+    @style_src.deleter
+    def style_src(self) -> None: ...
+    @property
+    def style_src_attr(self) -> Optional[str]: ...
+    @style_src_attr.setter
+    def style_src_attr(self, value: Optional[str]) -> None: ...
+    @style_src_attr.deleter
+    def style_src_attr(self) -> None: ...
+    @property
+    def style_src_elem(self) -> Optional[str]: ...
+    @style_src_elem.setter
+    def style_src_elem(self, value: Optional[str]) -> None: ...
+    @style_src_elem.deleter
+    def style_src_elem(self) -> None: ...
+    @property
+    def worker_src(self) -> Optional[str]: ...
+    @worker_src.setter
+    def worker_src(self, value: Optional[str]) -> None: ...
+    @worker_src.deleter
+    def worker_src(self) -> None: ...
+    provided: bool
+    def __init__(
+        self,
+        values: Union[Mapping[str, str], Iterable[Tuple[str, str]]] = (),
+        on_update: Optional[Callable[[ContentSecurityPolicy], None]] = None,
+    ) -> None: ...
+    def _get_value(self, key: str) -> Optional[str]: ...
+    def _set_value(self, key: str, value: str) -> None: ...
+    def _del_value(self, key: str) -> None: ...
+    def to_header(self) -> str: ...
